:art: move the render loop into gameState class
