+++ src/models/Game.ts
@@ -1,6 +1,7 @@
 import { Engine } from "@babylonjs/core";
 import { GameState } from "./GameState";
 import { InGameState } from "./scene/InGameState";
+import "@babylonjs/loaders/glTF";
 
 export class Game {
   public engine: Engine;
@@ -8,15 +9,8 @@ export class Game {
 
   constructor(canvas: HTMLCanvasElement) {
     this.engine = new Engine(canvas, true);
+    // Initialiser le premier état du jeu ici
     this.changeState(new InGameState(this, canvas));
-
-    // Initialiser la boucle de rendu ici
-    this.engine.runRenderLoop(() => {
-      if (this.currentState) {
-        this.currentState.update();
-        this.currentState.getScene().render();
-      }
-    }); // this.scene.render();
   }
 
   changeState(newState: GameState) {
